# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Humanity Ledger  Zero-Trust Analytics API
  version: 3.0.2
  description: |
    System Network Architecture REST API.
    All endpoints are protected by the TitaniumGate Zero-Trust Edge Middleware.
    Authentication requires EIP-191 ECDSA signatures and System Handshake JWTs.
  contact:
    name: Humanity Ledger Security Team
    email: security@humanityledger.io
servers:
  - url: https://api.humanityledger.io/v1
    description: Production Environment (European Region)
  - url: https://staging.humanityledger.io/v1
    description: Staging Environment (Beta)
  - url: http://localhost:3000/api
    description: Local Development
components:
  securitySchemes:
    SystemJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued after successful EIP-191 signature verification.
        Passed via HttpOnly cookie `system_handshake` or Authorization header.
    InternalHealthSecret:
      type: apiKey
      in: header
      name: X-Health-Secret
      description: Secret key for deep health checks (infrastructure probes only).
...
security:
  - SystemJWT: []
paths:
...
  /golden-ticket/verify:
    post:
      summary: Verify Golden Ticket ZK Proof
      description: Verifies Groth16 ZK-SNARK for tier assignment without exposing wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proof, publicSignals, walletAddress, timestamp]
              properties:
                proof:
                  type: object
                publicSignals:
                  type: object
                walletAddress:
                  type: string
                timestamp:
                  type: number
      responses:
        '200':
          description: ZK Proof verified, tier assigned
        '400':
          description: Invalid proof or nullifier already consumed
...